requirements.txt
Django==3.2.4
line-bot-sdk==1.19.0
python-decouple==v3.4